image creator/requirements.txt
google-generativeai>=0.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Pillow>=10.0.0
